Create upcasting
The provided Java program demonstrates the concept of method overriding and upcasting in object-oriented programming. The base class kumar has a method shoe() which prompts the user to enter a name and then prints it. The subclass sankar extends kumar and overrides the shoe() method, asking for a different input ("name1") and printing it. In the main method, an object of type kumar is instantiated as an object of sankar (upcasting). When obj.shoe() is called, the overridden method in the sankar class is executed, demonstrating polymorphism where the method called is determined by the actual object's type rather than the reference ty