Create  Build Array from Permutation
The buildArray method in the Solution class takes an integer array nums as input and constructs a new array ans of the same length. For each index i in the range of the array's length, the method sets ans[i] to nums[nums[i]], effectively rearranging the elements based on the values in nums. This transformation results in a permutation of the input array according to the indices specified by nums. The method returns the new array ans, demonstrating a basic but important concept in array manipulation and indexing in Java.