Create  Final Value of Variable After Performing Operations
The Solution class contains a method finalValueAfterOperations that takes an array of strings representing operations and returns the final value of a variable x after performing these operations. The operations are increment and decrement operations (++X, X++, --X, X--). The method initializes x to 0, iterates through each operation in the array, and updates x accordingly. For each increment operation, x is increased by 1, and for each decrement operation, x is decreased by 1. Finally, the method returns the resultant value of x, representing the cumulative effect of all operations.