Create  Minimum Number Game
This numberGame method in Java reverses the elements of the input array nums except for the last two elements, which it then swaps. The method starts by iterating over the first part of the array up to the second-to-last element, reversing the order of elements. Then, it explicitly swaps the last two elements. The purpose of this method appears to be to reverse the array while handling the last two elements differently to achieve a specific desired order.